Rename views and routes more consistently
