+++ app/routes/admin/email/CreateEmailListRoute.js
@@ -3,7 +3,7 @@
 import { compose } from 'redux';
 import { connect } from 'react-redux';
 import EmailListEditor from './components/EmailListEditor';
-import { createEmailList } from 'app/actions/emailListActions';
+import { createEmailList } from 'app/actions/EmailListActions';
 import { push } from 'react-router-redux';
 
 const mapDispatchToProps = { mutateFunction: createEmailList, push };
